tasks/booking_history_service/requirements.txt
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
kafka-python==2.0.2
